requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
loguru
joblib
pyyaml
tqdm
